Reimplement RBF transaction construction
This removes the mutually recursive construction logic with a modified
copy of the new send transaction construction loop. Several parts of the
logic are now shared via GATx functions. The actual construction loop
will be merged in the future once it can be generalised to cover send,
RBF and possibly elements.

This fixes:
- A bug identifying change outputs incorrectly when redepositing
with change (i.e. manual sends to the same subaccount)
- The lack of a summary dialog when no 2FA is present
- Respecting transaction limits to avoid 2FA for bumps under the users
spending limit

Some remaining issues:
- Entering incorrect 2FA requires the user to re-bump (this also affects
regular sending)
- The bump calculation is different and needs to be verified
- H/W signing needs to be tested

The above issues will be tackled after some further cleanup; this change
is already big enough.

+++ app/src/main/java/com/greenaddress/greenbits/ui/ReceiveFragment.java
@@ -303,7 +303,7 @@ public class ReceiveFragment extends SubaccountFragment implements OnDiscoveredT
             // TODO: non-fiat / non-assets values
             if (mAmountEdit.getText().toString().isEmpty())
                 return;
-            amount = (long) (Float.valueOf(mAmountEdit.getText().toString()).floatValue() * 100);
+            amount = (long) (Float.valueOf(mAmountEdit.getText().toString()) * 100);
         }
         mCurrentAddress = "";
         UI.disable(mCopyIcon);
